requirements.txt
gitpython
sklearn
pandas
numpy
unlzw
scipy
astropy
